requirements.txt
requests>=2.31.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
requests-oauthlib>=1.3.1
schedule>=1.2.0
pytest>=8.0.0
pytest-mock>=3.12.0
